# doc/api.yaml
openapi: 3.0.3
info:
  title: WASA-Exam
  description: API for the WASA Exam project.
  version: 1.0.0
security: []
...
paths:
...
  /users:
    get:
      tags:
        - Profile
      summary: Get all users
      description: Retrieve a list of all registered users.
      operationId: getAllUsers
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                description: Array of users
                minItems: 0
                maxItems: 1000
                items:
                  $ref: "#/components/schemas/User"
  /searchby:
    get:
      tags:
        - Profile
      summary: Search users
      description: Search for users by username.
      operationId: searchBy
      parameters:
        - name: user
          in: query
          description: Username to search for
          required: true
          schema:
            type: string
            description: Username query string
            minLength: 1
            maxLength: 50
            pattern: "^[a-zA-Z0-9_]+$"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of matching users
          content:
            application/json:
              schema:
                type: array
                description: Array of matching users
                minItems: 0
                maxItems: 100
                items:
                  $ref: "#/components/schemas/User"
  /user/username:
    put:
      tags:
        - Profile
      summary: Change username
      description: Update the authenticated user's username.
      operationId: setMyUserName
      security:
        - BearerAuth: []
      requestBody:
        description: New username
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Username"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Error
          content:
            application/json:
              schema:
                type: object
                description: Error message
                properties:
                  error:
                    type: string
                    description: Error detail
  /user/photo:
    put:
      tags:
        - Profile
      summary: Set profile photo
      description: Update the authenticated user's profile photo.
      operationId: setMyPhoto
      security:
        - BearerAuth: []
      requestBody:
        description: Photo data
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Photo payload
              required:
                - photo
              properties:
                photo:
                  type: string
                  description: Base64 encoded photo (or byte array)
                  format: byte
                  minLength: 1
                  maxLength: 5242880 # 5MB limit example
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Status message
...
  /conversations/{conversationId}/messages:
    parameters:
      - name: conversationId
        in: path
        required: true
        description: The ID of the conversation
        schema:
          type: string
          description: UUID of conversation
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
    post:
      tags:
        - Message
      summary: Send message
      description: Send a new message to the conversation.
      operationId: sendMessage
      security:
        - BearerAuth: []
      requestBody:
        description: Message content
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Message payload
              required:
                - content
              properties:
                content:
                  type: object
                  description: Content object
                  required:
                    - type
                    - value
                  properties:
                    type:
                      type: string
                      description: Type of content (text or photo)
                      enum: [text, photo]
                    value:
                      type: string
                      description: The text or base64 photo
                      minLength: 1
                      maxLength: 10000
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /conversations/{conversationId}/messages/{messageId}:
    parameters:
      - name: conversationId
        in: path
        required: true
        description: The ID of the conversation
        schema:
          type: string
          description: UUID of conversation
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
      - name: messageId
        in: path
        required: true
        description: The ID of the message
        schema:
          type: string
          description: UUID of message
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
    delete:
      tags:
        - Message
      summary: Delete message
      description: Delete a specific message.
      operationId: deleteMessage
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Status
                    example: "Deleted"
  /conversations/{conversationId}/messages/{messageId}/forward:
    parameters:
      - name: conversationId
        in: path
        required: true
        description: The ID of the conversation
        schema:
          type: string
          description: UUID of conversation
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
      - name: messageId
        in: path
        required: true
        description: The ID of the message
        schema:
          type: string
          description: UUID of message
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
    post:
      tags:
        - Message
      summary: Forward message
      description: Forward a message to another conversation.
      operationId: forwardMessage
      security:
        - BearerAuth: []
      requestBody:
        description: Target conversation
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Forward payload
              required:
                - targetConversationId
              properties:
                targetConversationId:
                  type: string
                  description: ID of target conversation
                  pattern: "^[0-9a-fA-F\\-]+$"
                  minLength: 36
                  maxLength: 36
      responses:
        "201":
          description: Forwarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /conversations/{conversationId}/messages/{messageId}/status:
    parameters:
      - name: conversationId
        in: path
        required: true
        description: The ID of the conversation
        schema:
          type: string
          description: UUID of conversation
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
      - name: messageId
        in: path
        required: true
        description: The ID of the message
        schema:
          type: string
          description: UUID of message
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
    post:
      tags:
        - Message
      summary: Set message status
      description: Update the read/delivered status of a message.
      operationId: setMessageStatus
      security:
        - BearerAuth: []
      requestBody:
        description: Status update
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Status payload
              required: [status]
              properties:
                status:
                  type: string
                  description: New status
                  enum: [delivered, read]
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Status
                    example: "Status updated"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Username:
      type: object
      description: Username schema
      required:
        - name
      properties:
        name:
          type: string
          description: The username
          minLength: 3
          maxLength: 20
          pattern: "^[a-zA-Z0-9_]+$"
          example: "Alex"
    User:
      type: object
      description: User profile info
      required:
        - id
        - username
      properties:
        id:
          type: string
          description: User UUID
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
          example: "00000000-0000-0000-0000-000000000001"
        username:
          $ref: "#/components/schemas/Username"
        profilePhoto:
          type: string
          description: Profile photo
          format: byte
          # nullable removed, use empty string or omission if needed
          minLength: 0
          maxLength: 5242880
          pattern: "^[A-Za-z0-9+/]*={0,2}$"
    Reaction:
      type: object
      description: Message reaction
      required:
        - type
        - authorId
      properties:
        type:
          type: string
          description: Type of reaction
          enum: [like, heart, laugh, sad_face, angry_face]
        authorId:
          type: string
          description: User ID who reacted
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
          example: "00000000-0000-0000-0000-000000000001"
    Message:
      type: object
      description: A message in a conversation
      required:
        - id
        - sender
        - content
        - timestamp
        - message_status
      properties:
        id:
          type: string
          description: Message UUID
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
          example: "00000000-0000-0000-0000-000000000001"
        sender:
          $ref: "#/components/schemas/User"
        content:
          type: object
          description: Message content
          required:
            - type
            - value
          properties:
            type:
              type: string
              description: Content type
              enum: [text, photo]
            value:
              type: string
              description: content value
              minLength: 1
              maxLength: 10000
              pattern: "^.*$"
        timestamp:
          type: string
          description: Message creation time
          format: date-time
          minLength: 20
          maxLength: 30
        message_status:
          type: string
          description: Status of the message
          enum: [sent, delivered, read]
        forwarded_from:
          type: string
          description: Original sender ID if forwarded
          pattern: "^[0-9a-fA-F\\-]+$"
          minLength: 36
          maxLength: 36
          # nullable: true
        reactions:
          type: array
          description: List of reactions
          minItems: 0
          maxItems: 100
          items:
            $ref: "#/components/schemas/Reaction"
          # nullable: true
...